enhanced: 기존 인게임 정보 조회, 티어 기준 랭킹 정보 조회 기능 개선

+++ src/main/java/onlysolorank/apiserver/domain/SummonerPlay.java
@@ -26,7 +26,7 @@ public class SummonerPlay {
     @Id
     private String id;
 
-    private Integer championId;
+    private Long championId;
     private String championName;
     private String puuid;
     private Integer totalPlays;
